updating simple socket server
